fix(condo): DOMA-8515 pr fixes
