+++ apps/condo/domains/meter/components/Meters/ChangeMeterStatusModal.tsx
@@ -1,13 +1,11 @@
-import { Row, Typography } from 'antd'
+import { Row } from 'antd'
 import { Gutter } from 'antd/lib/grid/row'
 import { CSSProperties } from 'react'
 
 import { useIntl } from '@open-condo/next/intl'
-import { Button, Modal } from '@open-condo/ui'
+import { Button, Modal, Typography } from '@open-condo/ui'
 
 import DatePicker from '@condo/domains/common/components/Pickers/DatePicker'
-
-const ARCHIVE_MODAL_TITLE_STYLES: CSSProperties = { fontSize: 16 }
 
 const ARCHIVE_MODAL_DATE_PICKER_STYLE: CSSProperties = { width: '100%' }
 const ARCHIVE_MODAL_DATE_PICKER_DATE_FORMAT = 'DD.MM.YYYY'
@@ -33,30 +31,33 @@ const ChangeMeterStatusModal = ({
     const ArchiveMeterButtonTextMessage = intl.formatMessage({ id: 'pages.condo.meter.Meter.PutToArchive.Modal.Text' })
     const CancelMessage = intl.formatMessage({ id: 'Cancel' })
     const ArchiveMeterButtonMessage = intl.formatMessage({ id: 'pages.condo.meter.Meter.putOutOfOrder' })
-    return (<Modal 
-        open={isShowStatusChangeModal}
-        title={ArchiveMeterModalTitleMessage}
-        onCancel={handleCloseStatusChangeModal}
-        footer={[
-            <Button key='cancel' type='secondary' onClick={handleCloseStatusChangeModal}>
-                {CancelMessage}
-            </Button>,
-            <Button key='submit' type='primary' onClick={changeMeterStatusToArchived} disabled={!selectedArchiveDate}>
-                {ArchiveMeterButtonMessage}
-            </Button>,
-        ]}
-    >
-        <Row gutter={ARCHIVE_MODAL_ROW_GUTTER}>
-            <Typography.Text style={ARCHIVE_MODAL_TITLE_STYLES} type='secondary'>
-                {ArchiveMeterButtonTextMessage}
-            </Typography.Text>
-            <DatePicker
-                format={ARCHIVE_MODAL_DATE_PICKER_DATE_FORMAT}
-                style={ARCHIVE_MODAL_DATE_PICKER_STYLE}
-                onChange={handleChangeSelectedArchiveDate}
-            />
-        </Row>
-    </Modal>)
+
+    return (
+        <Modal 
+            open={isShowStatusChangeModal}
+            title={ArchiveMeterModalTitleMessage}
+            onCancel={handleCloseStatusChangeModal}
+            footer={[
+                <Button key='cancel' type='secondary' onClick={handleCloseStatusChangeModal}>
+                    {CancelMessage}
+                </Button>,
+                <Button key='submit' type='primary' onClick={changeMeterStatusToArchived} disabled={!selectedArchiveDate}>
+                    {ArchiveMeterButtonMessage}
+                </Button>,
+            ]}
+        >
+            <Row gutter={ARCHIVE_MODAL_ROW_GUTTER}>
+                <Typography.Text type='secondary'>
+                    {ArchiveMeterButtonTextMessage}
+                </Typography.Text>
+                <DatePicker
+                    format={ARCHIVE_MODAL_DATE_PICKER_DATE_FORMAT}
+                    style={ARCHIVE_MODAL_DATE_PICKER_STYLE}
+                    onChange={handleChangeSelectedArchiveDate}
+                />
+            </Row>
+        </Modal>
+    )
 }
 
 export default ChangeMeterStatusModal
